requirements.txt
Flask==2.2.3
Flask-WTF==1.0.0
Werkzeug==2.2.3
pycryptodome==3.17.0
Pillow==9.4.0
numpy==1.24.2
zlib-wrapper==0.1.3
colorama==0.4.6
python-dotenv==1.0.0
flask-cors==3.0.10
requests==2.28.2
opencv-python==4.7.0.72
PyQt5==5.15.9 